Expose most of Gerrit's ChangeInfo in skylark
This allows accessing Gerrit's change information from
feedback workflows. I have not exposed complex objects
inside ChangeInfo for now.

Added a testing framework to evaluate skylark fields
(structField = true). This is important as it allows us
to verify the public interface that we are exposing to
users, instead of testing via the Java getters.

BUG=76115841
BUG=77858841
PiperOrigin-RevId: 193239659
Change-Id: I15c3d34b03a1ae4748b877272b34dbdaf02e8748