smart_prune: Diff that only returns affected files
Currently DiffUtil.diff returns the whole raw diff. This method exposes
a function that returns the affected files instead (relative to the
folders compared).

This function will be used to prune the affected changes for
CHANGE_REQUEST mode (if flag is set).

As part of this change I added a builder to make it easier to construct
different diff commands.

I also realized that we can use --no-index to compare folders inside
git directories. Added a TODO for now.

Change-Id: I1df997b44ca1bd89fe8efb4c9bad66c86f15e638