Share module instances between loaded files
We are instantiating a new module instance per file loaded. This is
generally K except for GlobalMigrations. Since it is truly global.

An alternative would be to have two levels of modules: GlobalMigrations
like ones and the rest. But we also have another module like this
internally, so it would be more complex.

We might revisit this in the future.

FIX=78567144

Change-Id: I87c1d4c6d4be46fb7680422abe8992ce7a14b26f