Move setting the currentConfigFile just before eval
After ef15e5bc, we were setting the current config file too early.
The reason is that after setting it, we were recursively evaluating
the dependencies. This means that we would keep the last dependencies
value instead of main (or the file that depends on the most inner
dep).

FIX=78567144
Change-Id: I4261cafb98c2313e9d41c150517d9a631ea2580e