Add --default-author flag
Default author needs to be set in the configuration. This is a
problem when the current running user wants to overwrite the
default to be they, since it requires changing the copy.bara.sky
file each time.

With this change, the user can pass a flag to overwrite the default
author.

We could have read it from $HOME/.gitconfig to make it more convenient,
but this would be less flexible, not all our users use git and would
make workflow depend on git.

Instead the user can create a wrapper script that does something like:

    copybara copy.bara.sky workflow --default-author "$(echo "$(git config user.name) <$(git config user.email)>")"

FIX=77914972

Change-Id: Ib1613fbe6100e1091a9dc4d32d6cf49664fc1d88