Use GNU Patch when possible
Currently we use 'git apply' for applying patches. There are cases
where patches don't apply that GNU Patch can patch successfully.

This change allows to use patch instead. For now the default is
still using 'git apply' until we make it available in the internal
service.

We also default to 'git apply' if the found version for patch is
too old or if the user uses excludes in the Skylark configuration.

The test logic is a bit convoluted since it has to work externally
in OSX and internally:

   - We skip some tests in OSX if version is too old, since the
     default version is too old. Mainly the ones that force to
     run with whatever patch version is found.

   - Change the path from 'patch' (uses PATH environment) to
     '/usr/bin/patch' if that version is too old. This solves an
     internal testing issue.

The command run is:

   patch -t -p <STRIP_NUM> [-R]

BUG=78536811
Change-Id: I1281cbabd904c97ceae934d8f82a72fb022f78bb