Freeze globals once a config file is loaded
Currently we are not freezing the globals from the Skylark file
dependencies. This causes an internal error if the user tries
to mutate the global from the dependency file. The reason is
that there is a locking system per environment and each config
file has its own environment.

Now we freeze all the globals so that they cannot be modified
after loaded. This has the effect of not allowign modifying
global variables from dynamic transformations/feedback. But
I consider that a feature. We don't want state between dynamic
transformations.

An alternative considered was to freeze the whole environment:

    env.mutability().freeze()

The problem with that is that Skylark forbids calls to functions
of that environment. This prevent us from calling dynamic
transformations once the config is loaded.

FIX=78648977
Change-Id: I33c9d6de039d7a286cabbe8afc36738a48c9fa8d