Fix another frezable exception issue
We were passing an immutable environment for executing dynamic
functions. Not we create a hierarchy of immutable environments
(Where each node is the file being loaded) and for each node
we create a mutable enviroment that depends on the environment
node and that it can be used for executing Skylark dynamic
actions.

We use a supplier so that everytime we call a top-level skylark
function (for example SkylarkTransformation), we create a new
mutable enviroment for that execution.

One not obvious thing is that that environment doesn't receive
any globals or extensions. Since they are already internally
in the function we call. Passing them again would make those
variables visible to multiple files (like globals).

Change-Id: Ic4cb4319a3a4a6eb4d3843305c70486bdc504edb