Use head from GitHub PR for integrate SHA-1
Currently we are setting the integrate SHA-1 to the temporary
merge SHA-1 if the user sets use_merge = True. This is incorrect
since we want to create a merge with HEAD of the PR.

Even worse, because that ref is a moving target (every commit
in the baseline changes the merge ref for the PR), it is
very likely that the SHA-1 is not reacheable from any head or
that it is GCed.

FIXED=79161317

Change-Id: I116b2fb29ef0e9a3dd7ba5e71dcf303a0345abca