Create ModuleSet
ModuleSet is a better scoped object that can be used for evaluating
Skylark. We still have some weirdness with the options but nothing
prevent us from reparsing them again if we need multiple evaluations
in the future.

The reason for this change is that I am moving the modules to be
non-static. So we'll need to pass them as normal global variables.
Instead of passing Options and module map (name -> object), we
pass the ModuleSet so that it is more difficult to get it wrong.

As part of this change I changed most of our tests to use an enhanced
SkylarkTestExecutor that has all the default modules for its
environment (OSS or internal), plus the testing one that is used
by severeal tests. We still can pass additional mock modules
but in general we should prefer relying in the ModuleSupplier, since
that is what real code is going to execute.

I also removed most if not all of the usages of SkylarkParser in favor
of the test executor. We already had all the functionality there and
it simplifies the usage.

PiperOrigin-RevId: 197153092
Change-Id: I8d70715b3cebac174d6bd8ce2cb424ce371fa944