Initial check infrastructure for feedback
Feedback actions are implemented in Skylark by users, and they
expose access to RPCs to code review systems like Gerrit or GitHub PRs.

For feedback migrations internal->OSS this could have the potential
risk of leaking confidential information.

The internal version of Copybara will prevent this by running an internal
tool that checks for leaks at the transport level. This is not part of the
OSS project but we still want to expose a mechanism to verify requests
before making RPC  calls.

This mechanism can be easily extended externally by implementing a
Checker.

BUG=79533809
PiperOrigin-RevId: 197621951
Change-Id: I8df45d309faea468c04ba0de99133a3e08fe7d7d