Use a graph for origin.changes() response
Currently Copybara returns changes as a flatten list. This works fine
for normal use cases since we navigate only the first parent of the
change, either because the VCS doesn't support graph of changes or
because we force the traversal.

The problem is that git.origin has an option to traverse all the
parents (first_parent = False). This is very useful for non-linear
history project since usually the merge is not very interesting. The
alternative of using include_branch_commit_logs has lot of issues (it
is just a text, cannot be use with squash_notes or others, don't
provide the labels, etc).

In some circumstances where the merge is a noop but some non-first
parent commits of the merge are a change and a revert, we were
detecting as the head change that one instead of more recent ones in
the main branch. The effect is that we revert or not include some
changes and we incorrectly generate the squash notes.

The solutions considered were:

   - Move all the changes computation to the origins. This makes sense
     since they know better than workflow how to deal with their
     history. But --read-config-from-change makes it impossible, as
     the decision of knowing if a change is included or not requires
     loading a workflow at that revision for getting the glob.

   - Read from head instead of the detected change. While this would
     create the correct file contents it won't generate a correct
     metadata.squash_notes.

   - Provide more information to the workflow about the relationship
     between the changes. The best representation of this is a
     graph.

The solution supports unconnected graph nodes for potential cases where
we cannot find a parent-child relationship. But we have a check to
forbid that. In case of an issue we can quickly remove that check.

The graph implementation include non-first parent changes iff the merge
commit is included. We define first-parent changes the nodes that don't
have children (tip or unconnected) or changes that are first-parent of
a first-parent.

For now we've only fixed the SQUASH case. ITERATIVE has a similar
problem, but first_parent = False is usually not used there (teams use
include_branch_commit_logs). But this is a more difficult problem to
fix. It can be easily fixable in a mode like include_branch_commit_logs
(semi SQUASH) but would be harder for pure ITERATIVE, since it requires
us to load many workflow at the same time in memory (merge + non-first
parents).

As part of the change we stop using the roots in git.origin for changes.
This fixes b/80493133 for the origin. roots are mean to be an
optimization but in some circumstances was used as 'correctness'. But
the main reason is that allows us to compute parent/child relationships
for the merges. Otherwise would have unconnected non-first parents to
the merges. And that would prevent us from detecting if we need to
include a change or not.

We maintain the filtering by roots in a couple of places:

   - git.origin visitChanges: It is common to have multiple projects
     that are exported/imported from one git repo (mono-repo). This
     allows us to not confuse folders of other migrations (Since
     they'll have a X-RevId). This is not ideal and I am not happy,
     but changing this behavior mean configurable label for rev id.

   - include_branch_commit_logs: This is deprecated but we used for
     massive imports. We don't want to break the current behavior until
     we replace. It is used internally for one off migrations.

On the other hand, git origin change() doesn't fail with
EmptyChangeException when it doesn't affect the roots. This is wrong
since origin_files could have changed. We use this in two cases:

   - User manually tell us the baseline in CHANGE_REQUEST: We trust
     the user here. And it will still generate an empty change error
     if really empty. If this is required in the future we should do
     the check in the workflow.

   - Validate the baseline with check_last_rev_state. If this is
     required we should do the check in workflow.

Several git.origin tests were changed to use workflow for validating
excluded changes, since we've move that logic to the workflow (we
already had it but we have removed from git.origin). Also that logic
was incomplete as we only used the roots.

Finally getChangesAsListForTest is an implementation that returns
a naive flatten list of the changes graph. The reason is to minimize
the already big change (we can change later if it bothers us).

PiperOrigin-RevId: 200397940
Change-Id: I59a008888cc16f9cbf89c87fafd2732a736037c4