git.gerrit_destination: Empty patchset detection
Implement empty patchset detection. If set to false, Copybara
will download the latest PatchSet, cherrypick it in top of the
new baseline and compare the tree with the commit to be submitted.
If the tree is the same it fails with EmptyChangeException (noop).

This is too complex because several times we work with a detached
HEAD. I plan to refactor this code to always use a branch at this
stage (b/111567027).

Change-Id: Ia67c3b7bf0c2e9098bf6fe3ad639b03da171e6c4