Reset work-tree always for git.destination
The special cases (rebase or local repo) are not worth to be special.
We can do always. This reduces the complexity of the gerrit empty
patchset detector. There was a problem where sometimes we don't left
files in the work-tree and git stash doesn't fail (but git stash
pop does if no stash references).

BUG=111314064
PiperOrigin-RevId: 205293176
Change-Id: Idb66cc23dcb731d4dc93a06c5b46f1f3bfb2e6f7