Honor --dry-run in git.mirror
Dry-run was only honored for core.workflow.

Moved the flag to GeneralOptions and honor the flag in git.mirror. We
assume that if we run in dryrun mode, we didn't push anything. An
alternative was to make GitMirrorOptions.mirror to return a boolean
but we never know for sure if mirror was a noop or not. We might
revisit this in the future.

FIXED=111751012
PiperOrigin-RevId: 205710081
Change-Id: Ib3099e8a47ee48b3ea428ac08340802c6da4f256