2nd attempt: folder.origin symlinks fix
If we have a folder.origin with a glob like:

    origin_files = glob([
        "foo/**",
	"bar/**",
    ])

And foo contains a symlink to a file in bar, we fail saying that
the the symlink escaped the roots. This is incorrect and should
be allowed.

This change fixes the issue by using the glob instead to see if
the symlinks escape. I also took the opportunity to simplify the
symlink checking code.

As part of this change, and in order to support the old behavior,
we allow foo/symlink -> '.' like symlinks (that resolve to foo).
Technically the glob foo/** doesn't include it but this would be
confusing for our users. Specially for the default origin files
["**"], that wouldn't allow this and is impossible for our users
to fix (except by excluding the symlink).

What changed since last attempt 119abce4:

  - Fix a bug for non-normalized from/to paths. Pathmatchers are
    basically regexes, so a glob like glob:foo/bar/file* won't
    match foo/bar/../bar/file1, even if it represents a path
    (if normalized) that would match. Instead we normalize from,
    to and files so that we don't get surprises.

    This issue can happen if someone declares a folder.origin
    with a glob like glob(["foo/file*"]) and passes as the reference
    (the folder) '..' in the CLI.

  - Added a toString() for UnionGlob so that errors/info
    shows properly.

FIXED=113570475

Change-Id: Ib06d81b1d25bdc8f3d8035d9268b0c19d7d38615