Make default event monitor print events to the console
Messages are only printed if --verbose is used.

We were already doing this for some events in the default monitor,
this change is extending it to every method.

This can be very useful to debug issues, as we'll be able to see all
the events and specially the destination effects.

BUG=76112237
BUG=113288236
PiperOrigin-RevId: 212819698
Change-Id: I795a026796bd3316836037385289936a3bc66098