Use explicit values in new_destination_ref
Do not infer from the endpoint the type and
url if they are not set.

Type is now a mandatory field.

DestinationRef can have a null url, so if
it's not provided by the user it will be
null.

This is aligned with how we create DestinationRefs
natively in Copybara. Using the values from
Endpoint#describe() didn't generate meaningful
values.

BUG=76112237
BUG=113288236
PiperOrigin-RevId: 213639068
Change-Id: Ie3d4feca21b3ccc64de6507e50d28b2f3770d3d8