Replace changes graph with conditional changes
The graph solution had a fundamental problem since it didn't include
all the parents of the git repo. That means that we would identify
incorrectly as first parent, changes that weren't first parents.

In particular, this case:

    * <-- merge that doesn't affect master
    |\
    | * <-- feature: detected as affected
    * | <-- last_rev
    |/
    *
would incorrectly detect the branch change as first-parent and
proceed to migrate it. Instead, the correct thing to do is to
fail with noop, since there hasn't been changes in master
since the last import.

The new solution computes the changes as a list (as we did
before) but also computes a list of changes that should only
be included if other changes are included. This allows us to
say that "feature" should only be included if the merge is
included. The difference is that inside git.origin we have
all the information for computing that.

As part of this change I also added parents information to
changes (optional).

An alternative solution is to compute the conditional changes
at the workflow level. The problem is that we might have
changes with the parent field filled but that are not connected
to previous included changes (for example for our internal
piper implementation where baseline could be used as parent
but it is not going to create a connected tree).

BUG=117308901
PiperOrigin-RevId: 216799529
Change-Id: I8fa2da8b5a9583aab8df49e61db5fac3be5c54f8