Run after_migration hooks after a non-user error is thrown
We were not running the hook for non-user errors (RepoException,
IOException, etc), but there is no reason for that. The Skylark
action can decide what to do depending on the type of error.

As part of this change I'm introducing a new type of error in
the migration effect type: TEMPORARY_ERROR.

PiperOrigin-RevId: 216860908
Change-Id: Id7870ea91bb991204e75179da7b0b81f4acd6b2f