Introduce flag --nogit-prompt
This flag sets Git's environment variable GIT_TERMINAL_PROMPT=0
so Git won't ask for username/password when it doesn't have
the credentials for a repo.

This is useful mainly for automated jobs running Copybara, like
our internal service or a script.

As part of this change I've introduced a GitEnvironment class
that encapsulates the logic of these Git environment variables
and replaces Map<String, String> by a strong type. I've also
consolidated the logic of GitExecPath into GitEnvironment.

We still can improve how the environment is used from tests,
but not in this change.

PiperOrigin-RevId: 219854339
Change-Id: I9190f2c4d6dd0262736bda127248b223f914cc1d