Fix 'branch already exists' for git.destination
When dry run mode is used, we were creating many writers per
migration when we couldn't detect the changes. The effect of this is
that we would try to recreate a branch when --git-destination-path
flag was used, and that would fail in git.

This change refactor the code to always use the same writer except when
check_last_rev_state is used. This one needs a new writer so that the
branch that uses for validation is different. I hope we can get rid
of check_last_rev_state soon too.

This change is part of the effort of removing skip_push from
git.destination.

FIXED=119526684
PiperOrigin-RevId: 223196365
Change-Id: I4e411dafa9a58df638f46b2349ccf0b080fdd8f7