Use Starlark http_archive function
This removes the temporary flag for using the deprecated native flag
and replaces it with the Starlkark version.

As part of this change, I am also removing Cram since it is not being
used and it uses new_http_archive (That might or might not work
with the Starlark library).

Change-Id: I0b60ce4f3d2101522249c9c59299f3f1affb8772