Detect non-zero exit status due to timeout kills
Currently when we kill commands due to timeout, we don't show any
information that this was due to a timeout and not a failure of the
command. This confuses users as it is not evident the reason was
a timeout (For example imagine a fetch from a repo that doesn't show
any output and then gets killed by timeout).

With this change we detect the timeout in two classes of exception:

  - AbnormalTerminationException: This happens for normal commands
    that end with exit code 129. Shell library intercept this and
    uses this exception.

  - BadExitStatusException: This happens for commands that capture
    the signal and exit with a custom exit code (For example hg).

BUG=120030093
PiperOrigin-RevId: 227171737
Change-Id: Ia4755126f92944f5da95cc2220e13b56b2952efb