Allow no changes in include_branch_commit_logs
Currently if include_branch_commit_logs is used but the merged commits
(non-first parent changes) are for files not in the included roots,
we fail with a IndexOutOfBoundsException exception. The reason is
that we don't find any commit in the history of the non-first parent
that matches the roots.

This change simply skip when there are no commits. This is not the
proper fix, as include_branch_commit_logs shouldn't filter by roots
(glob filtering is much more complicated, and is done at the workflow
level). But it is good enough since the field is deprecated in favor
of first_parent = False (We should add support for ITERATIVE mode).

FIXED=122489022
Change-Id: I099d91fcafd6280f05e5e3bbe040ec4848192575