Use complete refs for GitHubWriterHook
Our internal implementation requires that we use full refspecs when
pushing/fetching (refs/heads/master vs master). GitHubWriterHook is
using short branch name instead of the complete ref.

Extracted the validator to a common class and added to a our GitHub
specific origin/destinations (generic git.origin/destination could
potentially be affected but it is not easy as it is to change the
tests and the GH ones give enough coverage for the base classes).

Also removed some non-used code and replaced an assertion with the
enum.

FIXED=70173346
Change-Id: I5c5d746462174e8c225da65868df7d747045ac0a