hg.origin: Be able to fetch non-default refs
Currently if hg.origin is used with non-default revs (default, tip,
etc. are default revs), it fails since it doesn't fetch the
tag/branch name.

IOW. If we use example-1.0.1 as the ref, it will pull it but
won't create tag example-1.0.1. Note that it will create
previous commits ones (example-1.0, example-0.9, etc.). This
makes our next call to 'hg identify example-1.0.1' fail, as what
we pulled is only in 'tip'.

This changes pulls all the refs so that we get the tag created.
This is also faster for most of the repos and the size difference
usually is small. The reason is that the default refs is usually
precomputed as a package in the server.

Finally, as part of the change, I introduced an optimization to not
pull SHA-1 if they are already in our local repo. This makes it
a bit faster for checking resolving last-rev.

Side change: Fixed documentation formatting.

FIXED=122907888

Change-Id: Ief748529d1faf8839f627fc7d58379a69993c43e